move page options menu into widgets toolkit
ahead of sharing it with the detail and report pages
