+++ projects/packages/premium-analytics/packages/widgets-toolkit/package.json
@@ -11,6 +11,7 @@
 		"*.scss"
 	],
 	"dependencies": {
+		"@automattic/jetpack-analytics": "workspace:*",
 		"@automattic/jetpack-script-data": "workspace:*",
 		"@jetpack-premium-analytics/data": "workspace:*",
 		"@jetpack-premium-analytics/datetime": "workspace:*",
