+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/components/index.ts
@@ -152,6 +152,7 @@ export {
 	type DetailPageSectionProps,
 	type DetailPageShellProps,
 } from './detail-page';
+export { PageOptionsMenu } from './page-options-menu';
 export {
 	ReportCsvDownloadButton,
 	type ReportCsvDownloadButtonProps,
