+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/components/page-options-menu/index.ts
@@ -1,0 +1,1 @@
+export { PageOptionsMenu } from './page-options-menu';
