+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/components/page-options-menu/return-to-classic-stats.ts
