+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/index.ts
@@ -110,6 +110,7 @@ export {
 	type DetailPageLayoutProps,
 	type DetailPageSectionProps,
 	type DetailPageShellProps,
+	PageOptionsMenu,
 	LocationsGeoChart,
 	type LocationsGeoChartProps,
 	type LocationsGeoFocusCountry,
@@ -243,6 +244,7 @@ export {
 	useSegmentStyles,
 	useSeriesStyles,
 	useStoredDetailLayout,
+	useTrackEvent,
 	useViewportWidth,
 	useWidgetDrillDown,
 } from './hooks';
