+++ projects/packages/premium-analytics/routes/dashboard/components/index.ts
@@ -1,4 +1,3 @@
-export { DashboardOptionsMenu } from './options-menu/dashboard-options-menu';
 export { DashboardSections } from './dashboard-sections/dashboard-sections';
 export { OnboardingTour, onboardingTourSteps } from './onboarding-tour';
 export { RefreshFailureNotice } from './refresh-failure-notice/refresh-failure-notice';
